+++ 1Week_Mission/src/main/java/com/ll/exam/mutbooks/app/post/form/dto/PostDto.java
@@ -1,4 +1,4 @@
-package com.ll.exam.mutbooks.app.post.form;
+package com.ll.exam.mutbooks.app.post.form.dto;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -7,7 +7,7 @@ import javax.validation.constraints.NotEmpty;
 
 @Setter
 @Getter
-public class PostForm {
+public class PostDto {
     @NotEmpty
     private String subject;
     @NotEmpty
